% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ces_list_methods}
\alias{ces_list_methods}
\title{List registered disaggregation methods}
\usage{
ces_list_methods()
}
\value{
Data frame with columns name, type, description.
}
\description{
List registered disaggregation methods
}
